requirements.txt
PyQt5>=5.15.0
pyserial>=3.5
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
python-usbtmc>=0.0.3
PyQt5-sip>=12.9.0
PyQtChart>=5.15.0
pyusb>=1.2.1